+++ src/App/Documentation/utilities/Display/index.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { ComponentPreview, DocToc } from "#";
+import { ComponentPreview, DocToc, Property } from "#";
 
 const HowItWorks = () => (
     <>
         <h2 id="how-it-works">How it works</h2>
         <p>
-            Change the value of the <a href="https://developer.mozilla.org/en-US/docs/Web/CSS/display">display property</a> with our responsive display utility classes. We purposely support only a subset of all possible values for <code className="token property">display</code>. Classes can be combined for various effects as you need.
+            Change the value of the <a href="https://developer.mozilla.org/en-US/docs/Web/CSS/display">display property</a> with our responsive display utility classes. We purposely support only a subset of all possible values for <Property value="display" />. Classes can be combined for various effects as you need.
         </p>
     </>
 );
@@ -16,30 +16,30 @@ const Notation = () => (
     <>
         <h2 id="notation">Notation</h2>
         <p>
-            Display utility classes that apply to all <Link to="/docs/core/breakpoints">breakpoints</Link>, from <code className="token property">xs</code> to <code className="token property">xxl</code>, have no breakpoint abbreviation in them. This is because those classes are applied from <code className="token property">min-width: 0;</code> and up, and thus are not bound by a media query. The remaining breakpoints, however, do include a breakpoint abbreviation.
+            Display utility classes that apply to all <Link to="/docs/core/breakpoints">breakpoints</Link>, from <Property value="xs" /> to <Property value="xxl" />, have no breakpoint abbreviation in them. This is because those classes are applied from <Property value="min-width: 0;" /> and up, and thus are not bound by a media query. The remaining breakpoints, however, do include a breakpoint abbreviation.
         </p>
         <p>As such, the classes are named using the format:</p>
         <ul>
             <li>
-                <code className="token property">{".d-{value}"}</code> for <code className="token property">xs</code>
+                <Property value=".d-{value}" /> for <Property value="xs" />
             </li>
             <li>
-                <code className="token property">{".d-{breakpoint}-{value}"}</code> for <code className="token property">sm</code>, <code className="token property">md</code>, <code className="token property">lg</code>, <code className="token property">xl</code> and <code className="token property">xxl</code>.
+                <Property value=".d-{breakpoint}-{value}" /> for <Property value="sm" />, <Property value="md" />, <Property value="lg" />, <Property value="xl" /> and <Property value="xxl" />.
             </li>
         </ul>
         <p>Where <i>value</i> is one of:</p>
         <ul>
-            <li><code className="token property">none</code></li>
-            <li><code className="token property">inline</code></li>
-            <li><code className="token property">inline-block</code></li>
-            <li><code className="token property">block</code></li>
-            <li><code className="token property">table</code></li>
-            <li><code className="token property">table-cell</code></li>
-            <li><code className="token property">table-row</code></li>
-            <li><code className="token property">flex</code></li>
-            <li><code className="token property">inline-flex</code></li>
+            <li><Property value="none" /></li>
+            <li><Property value="inline" /></li>
+            <li><Property value="inline-block" /></li>
+            <li><Property value="block" /></li>
+            <li><Property value="table" /></li>
+            <li><Property value="table-cell" /></li>
+            <li><Property value="table-row" /></li>
+            <li><Property value="flex" /></li>
+            <li><Property value="inline-flex" /></li>
         </ul>
-        <p>The media queries effect screen widths with the given breakpoint <i>or larger</i>. For example, <code className="token property">.d-xl-none</code> sets <code className="token property">display: none;</code> on both <code className="token property">xl</code> and <code className="token property">xxl</code> screens.</p>
+        <p>The media queries effect screen widths with the given breakpoint <i>or larger</i>. For example, <Property value=".d-xl-none" /> sets <Property value="display: none;" /> on both <Property value="xl" /> and <Property value="xxl" /> screens.</p>
     </>
 );
 
@@ -64,10 +64,10 @@ const HidingElements = () => (
             For faster mobile-friendly development, use responsive display classes for showing and hiding elements by device. Avoid creating entirely different versions of the same site, instead hide element responsively for each screen size.
         </p>
         <p>
-            To hide elements simply use the <code className="token property">.d-none</code> class or one of the <code className="token property">{".d-{sm,md,lg,xl}-none"}</code> classes for any responsive screen variation.
+            To hide elements simply use the <Property value=".d-none" /> class or one of the <Property value=".d-{sm,md,lg,xl}-none" /> classes for any responsive screen variation.
         </p>
         <p>
-            To show an element only on a given interval of screen sizes you can combine one <code className="token property">.d-*-none</code> class with a <code className="token property">.d-*-*</code> class, for example <code className="token property">.d-none .d-md-block .d-xl-none</code> will hide the element for all screen sizes except on medium and large devices.
+            To show an element only on a given interval of screen sizes you can combine one <Property value=".d-*-none" /> class with a <Property value=".d-*-*" /> class, for example <Property value=".d-none .d-md-block .d-xl-none" /> will hide the element for all screen sizes except on medium and large devices.
         </p>
         <table className="table table-striped">
             <thead>
@@ -79,59 +79,59 @@ const HidingElements = () => (
             <tbody>
                 <tr>
                     <td>Hidden on all</td>
-                    <td><code className="token property">.d-none</code></td>
+                    <td><Property value=".d-none" /></td>
                 </tr>
                 <tr>
                     <td>Hidden only on xs</td>
-                    <td><code className="token property">.d-none .d-sm-block</code></td>
+                    <td><Property value=".d-none .d-sm-block" /></td>
                 </tr>
                 <tr>
                     <td>Hidden only on sm</td>
-                    <td><code className="token property">.d-sm-none .d-md-block</code></td>
+                    <td><Property value=".d-sm-none .d-md-block" /></td>
                 </tr>
                 <tr>
                     <td>Hidden only on md</td>
-                    <td><code className="token property">.d-md-none .d-lg-block</code></td>
+                    <td><Property value=".d-md-none .d-lg-block" /></td>
                 </tr>
                 <tr>
                     <td>Hidden only on lg</td>
-                    <td><code className="token property">.d-lg-none .d-xl-block</code></td>
+                    <td><Property value=".d-lg-none .d-xl-block" /></td>
                 </tr>
                 <tr>
                     <td>Hidden only on xl</td>
-                    <td><code className="token property">.d-xl-none .d-xxl-block</code></td>
+                    <td><Property value=".d-xl-none .d-xxl-block" /></td>
                 </tr>
                 <tr>
                     <td>Hidden only on xxl</td>
-                    <td><code className="token property">.d-xxl-none</code></td>
+                    <td><Property value=".d-xxl-none" /></td>
                 </tr>
                 <tr>
                     <td>Visible on all</td>
-                    <td><code className="token property">.d-block</code></td>
+                    <td><Property value=".d-block" /></td>
                 </tr>
                 <tr>
                     <td>Visible only on xs</td>
-                    <td><code className="token property">.d-block .d-sm-none</code></td>
+                    <td><Property value=".d-block .d-sm-none" /></td>
                 </tr>
                 <tr>
                     <td>Visible only on sm</td>
-                    <td><code className="token property">.d-sm-block .d-md-none</code></td>
+                    <td><Property value=".d-sm-block .d-md-none" /></td>
                 </tr>
                 <tr>
                     <td>Visible only on md</td>
-                    <td><code className="token property">.d-md-block .d-lg-none</code></td>
+                    <td><Property value=".d-md-block .d-lg-none" /></td>
                 </tr>
                 <tr>
                     <td>Visible only on lg</td>
-                    <td><code className="token property">.d-lg-block .d-xl-none</code></td>
+                    <td><Property value=".d-lg-block .d-xl-none" /></td>
                 </tr>
                 <tr>
                     <td>Visible only on xl</td>
-                    <td><code className="token property">.d-xl-block .d-xxl-none</code></td>
+                    <td><Property value=".d-xl-block .d-xxl-none" /></td>
                 </tr>
                 <tr>
                     <td>Visible only on xxl</td>
-                    <td><code className="token property">.d-xxl-block</code></td>
+                    <td><Property value=".d-xxl-block" /></td>
                 </tr>
             </tbody>
         </table>