+++ src/px-script/index.js
@@ -6,6 +6,7 @@ import button from "./button";
 import datepicker from "./datepicker";
 import dialog from "./dialog";
 import loader from "./loader";
+import nav from "./nav";
 import rangeslider from "./rangeslider";
 import sheet from "./sheet";
 import script from "./script";
@@ -27,6 +28,7 @@ const px = {
     datepicker,
     dialog,
     loader,
+    nav,
     rangeslider,
     sheet,
     script,
@@ -47,4 +49,4 @@ if (!window.stopPx) {
 
 export default px;
 
-export { actionList, alert, button, datepicker, dialog, loader, rangeslider, sheet, script, tabs, toast, topbar, utils, validation };
+export { actionList, alert, button, datepicker, dialog, loader, rangeslider, sheet, nav, script, tabs, toast, topbar, utils, validation };