+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+- Added Sidebar component.
+- Added documentation section for text utlities.
+- Added documentation section for flex utlities.
+- Added documentation section for sizing utlities.
 - Added functionality to programmatically open and close dialog component.
 - Added functionality to programmatically open and close sheet component.
 
@@ -17,3 +21,5 @@
 - Fixed bug where sheet component adds padding-right when no scrollbar exists
 
 ### Removed
+
+- Removed deprecated progress tracker component.